Restrict 'polarization' to 's' and 'p' OR rightHanded OR left-handed.
Restrict 'polarization' to the cases which are currently relevant for measurements. I have deleted 'shiftedWavelengths' and singleShifted.json, because this is a very rare case which needs to be developed to a higher quality at a later stage. Now, wavelength and polarization are defined within 'incidence' only. I have created an example and valid and invalid tests for polarization. Relates to #48