Restrict `symmetries`, fix `emergenceDirection`
I have updated the description of `symmetry` in `opticalData` and extended it by `mirrorAndRotational`. I removed the `array` to forbid combinations because `azimuthalAngleInvariance`, `profileAngle`, `mirror`, `rotationalDependingOnAzimuth`, `mirrorAndRotational` should cover everything and without overlap.

I searched the repository for `incidence angle`, `incidence` and `incidence-angle` and updated the descriptions where necessary.

I have added a `oneOf` to the `angleDependency` of `photovoltaicData` and updated the descriptions and `bipvModuleBuildingNatISE`.

I have deleted `reflectance` and `nearnormal` from `emergenceDirection` to reflect the updates in `incidenceDirection`.