Add roughness and clean-up geometry.json
I have deleted `centerOfGravity`, because it is only helpful in a CAD
software.

I have deleted `bodies`, because this can be covered by `assembly`.

I have moved `definitionOfSurfacesAndPrimeDirection` from
opticalData.json to component.json#/$defs/characteristics, because it's
needed to define the roughness and possibly other properties beyond the
scope of opticalData.json.

I have added roughness/root-mean-square to component.json, because it
may be relevant beyond the scope of opticalData.json.

I moved the definition of `surface` to component.json because it is
needed there and can be referenced by opticalData.json.

Closes #204