Improve the structure of calorimetricData.json
Instead of one array with large items for example for `exterior`, there
is now only one object `fluxes` with arrays for `convective`,
`irradiance`, `fluid` and `electricity`. Only for `convective` and
`irradiance`, `side` can be defined. I deleted the option `other` in
`side`, because `side` is optional. Now it should be easier to define
for example several directions of irradiance and to parse them.