Remove `validity` except from `method`
I have deleted `validity` in calorimetric, optical and hygrothermal,
because the validity depends on the method. When a new version of the
method is published, the old data is still valid according to the old
version. validity/to doesn't help in this case. validity/from doesn't
help neither, because the properties of a component do not change and
are therefore always valid for this component.

I have deleted `validity` in common/approval, because there could be
ambigous meanings if there is an approval with a timestamp after
validity/to. If there is an approval, it's also valid for the past of
this component. If the databases do not return a current approval,
then it was withdrawn. If an approver wants to limit the duration of
the validity, she can do so in the `statement` of the approval which
we still need to implement.

I keep `validity` only in `method`, because methods can depend on
regulations or standards which change sometimes and deprecated methods
should still be available. Deprecated data does not need to be
available, because everyone can store it for themselves.

Closes #174