Offer detailed definition of calorimetric fluxes
I have added heat exchange by conduction, by long-wave heat transfer, by combination of conduction, convection and long-wave heat transfer and by mixed heat transfer.

With `primeSurface`, there is now a way to define the orientation of the installed sample. This is needed to retrieve the emittance of the sample.

I have deleted characteristics/coordinateSystem, because it is not needed anymore. The definition of the prime surface and prime direction in `component` is sufficient to define the direction of the irradiance.

I have updated the examples and added tests.