Correct the identifiers
The definition of `decentral` was lacking `additionalProperties`: false
and `required`: [`value`,`issuedBy`].

The tests uuidWithHyphen and uuidWithoutHyphen must contain only a
string. `central` is not a property but a definition.

smallGeneralComponentWithoutHyphen.json was correctly detected as
invalid.

The smallestValidJson cannot be an empty object anymore. The id must be
a UUID.

The identifier of koester, assemblyWhichIsPartOfAnotherAssembly and geometricAssemblyWhichIsPartOfAnotherAssembly where corrected.