Add `formatId` to the PropositionInputs
I have added the `formatId` of the dataFormat to `DataPropositionInput`,
`OpticalDataPropositionInput`, `CalorimetricDataPropositionInput`,
`HygrothermalDataPropositionInput` and
`PhotovoltaicDataPropositionInput`. I didn't add it to `interface Data`,
because the `formatId` is included in the `resources`.

I have also rearranged the order of the PropositionInputs to
`componentId`, `and`, `not`, `or`, `formatId`, `gValue`, `uValue`,`nearnormalHemisphericalVisibleTransmittance`,
`nearnormalHemisphericalVisibleReflectance`,
`nearnormalHemisphericalSolarTransmittance`,
`nearnormalHemisphericalSolarReflectance`, `infraredEmittance`.