Correct new examples, add "EN" to standardizers
I have corrected the two new examples testlabVE210609042606integral.json
and testlabVE210609042606spectral.json .

In addition to "CEN" and "CENELEC", I have added "EN" and "ETSI". "EN"
is important for practical reasons, because the users sometimes don't
know which European Standardization Organization has ratified the EN
standard. "CEN", "CENELEC" and "ETSI" are important for drafts which
are not (yet) ratified as EN standard.