Update metabase.graphql and example of assembly
The first version of the metabase.graphql was generated in an automated
way. It was not a valid GraphQL file. In order to change that, I have
added the scalars `Schema`, `Uuid` and `Cursor`. I also added `PageInfo`
to many `Connections`.

I have updated the examples and added the example
collectorFieldAsGeometricAssembly.json . I validated and formated the
schemas, examples and tests.

Closes #252